requirements.txt
tweepy
requests
replicate
python-dotenv
